requirements.txt
Scrapy==2.14.1
pytest-playwright==0.7.2
